+++ libsolidity/analysis/ControlFlowGraph.h
@@ -98,6 +98,8 @@ struct CFGNode
 	std::vector<CFGNode*> entries;
 	/// Exit nodes. All CFG nodes to which control flow may continue after this node.
 	std::vector<CFGNode*> exits;
+	/// Function calls done by this node
+	std::vector<FunctionCall const*> functionCalls;
 
 	/// Variable occurrences in the node.
 	std::vector<VariableOccurrence> variableOccurrences;
@@ -118,7 +120,7 @@ struct FunctionFlow
 	/// (e.g. all return statements of the function).
 	CFGNode* exit = nullptr;
 	/// Revert node. Control flow of the function in case of revert.
-	/// This node is empty does not have any exits, but may have multiple entries
+	/// This node is empty and does not have any exits, but may have multiple entries
 	/// (e.g. all assert, require, revert and throw statements).
 	CFGNode* revert = nullptr;
 	/// Transaction return node. Destination node for inline assembly "return" calls.
@@ -135,8 +137,13 @@ public:
 	bool constructFlow(ASTNode const& _astRoot);
 
 	bool visit(FunctionDefinition const& _function) override;
+	bool visit(ContractDefinition const& _contract) override;
 
-	FunctionFlow const& functionFlow(FunctionDefinition const& _function) const;
+	/// Create a function flow for the given function, using `_contract` as the
+	/// most derived contract
+	/// @param _function function to create the function flow for
+	/// @param _contract most derived contract or nullptr for free functions
+	FunctionFlow const& functionFlow(FunctionDefinition const& _function, ContractDefinition const* _contract = nullptr) const;
 
 	class NodeContainer
 	{
@@ -153,7 +160,7 @@ private:
 	/// are owned by the CFG class and stored in this container.
 	NodeContainer m_nodeContainer;
 
-	std::map<FunctionDefinition const*, std::unique_ptr<FunctionFlow>> m_functionControlFlow;
+	std::map<std::pair<ContractDefinition const*, FunctionDefinition const*>, std::unique_ptr<FunctionFlow>> m_functionControlFlow;
 };
 
 }